requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
seaborn>=0.11.0
matplotlib>=3.4.0
mlxtend>=0.19.0
fpdf2>=2.5.0
